Refactor prompt tests structure
